requirements.txt
folium==0.11.0
numpy==1.18.1
matplotlib==3.1.3
streamlit_folium==0.0.1
pandas==1.0.1
seaborn==0.10.0
streamlit==0.64.0
altair==4.1.0
joblib==0.14.1
Pillow==7.2.0
scikit_learn==0.23.1
